Split out the search view into a list and a search view, since only half the functionality will be needed in some instances
